requirements.txt
yfinance>=0.2.36
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
newsapi-python>=0.2.7
vaderSentiment>=3.3.2
matplotlib>=3.7.0
streamlit>=1.30.0 